#124 Minor refactoring separating yargs options and logic for adding options to execution
